moving azure clip storage to its own file
